Separate standalone notebook embed path and make non-deferred for now
